Implement kubectl port-forwarding as output
After creation of a service in k8s one often wants to configure it
directoy using the api of the service. However services are usually
not directly accessible after they are launched. Therefore add
ensure_port_forward which takes a local port and the target where to
forward to. It calls kubectl port-forward and returns the local port
which can be used by subsequent pulumi providers.